# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "User Service API",
    "version": "0.3.1"
  },
  "paths": {
    "/health": {
      "get": {
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "enum": ["ok"] },
                    "version": {
                      "type": "string",
                      "description": "Service semver from pyproject at install/build time."
                    }
                  },
                  "required": ["status", "version"]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "responses": {
          "200": {
            "description": "User list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserListResponse" }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a registry user (e.g. patient enrollment)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserCreateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing user updated (uuid upsert)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/api/v1/tenants/{tenant_uuid}/users": {
      "get": {
        "summary": "List users in a tenant",
        "parameters": [
          { "name": "tenant_uuid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100 } },
          { "name": "q", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Tenant-scoped user list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TenantUserListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_uuid}": {
      "get": {
        "parameters": [
          { "name": "user_uuid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "patch": {
        "parameters": [
          { "name": "user_uuid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user profile",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "403": {
            "description": "Cedar policy denied the requested field changes (e.g. self-service roles or identity fields)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Provision or refresh a user registry row from Authentication",
        "parameters": [
          { "name": "user_uuid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserProvisionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Existing row refreshed; roles unchanged",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "201": {
            "description": "Registry row created with roles: [] (identity only; assign roles via authorized create/update)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "403": {
            "description": "Only the authentication service token may call this route",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{user_uuid}/audits": {
      "get": {
        "parameters": [
          { "name": "user_uuid", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "User audit history",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/UserAuditListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/roles": {
      "get": {
        "responses": {
          "200": {
            "description": "Role catalog",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RoleCatalog" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "description": "Same as Authentication users.uuid (JWT sub)"
          },
          "tenant_uuid": { "type": "string", "format": "uuid" },
          "idp_id": { "type": "string", "minLength": 1 },
          "display_code": { "type": "string", "nullable": true },
          "first_name": { "type": "string", "nullable": true },
          "last_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "roles": { "type": "array", "items": { "type": "string" } },
          "tos_accepted": {
            "type": "boolean",
            "description": "True after the user accepts the platform Terms of Service in CDP UI"
          }
        },
        "required": ["uuid", "tenant_uuid", "idp_id", "roles", "tos_accepted"]
      },
      "UserListResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" }
        },
        "required": ["items", "total", "page", "page_size"]
      },
      "TenantUserListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/UserListResponse" },
          {
            "type": "object",
            "properties": {
              "tenant_uuid": { "type": "string", "format": "uuid" }
            },
            "required": ["tenant_uuid"]
          }
        ]
      },
      "UserUpdateRequest": {
        "type": "object",
        "properties": {
          "uuid": {
            "type": "string",
            "format": "uuid",
            "description": "Registry primary key. Accepted on PATCH for Cedar authorization only; self-service updates are forbidden (403). Not applied by PATCH handlers (use PUT provision for identity sync)."
          },
          "tenant_uuid": { "type": "string", "format": "uuid" },
          "idp_id": {
            "type": "string",
            "minLength": 1,
            "description": "Identity provider subject. Accepted on PATCH for Cedar authorization only; self-service updates are forbidden (403). Not applied by PATCH handlers (use PUT provision)."
          },
          "actor_class": {
            "type": "string",
            "enum": ["operator", "study", "clinician", "patient"],
            "description": "Tier-1 actor class. Accepted on PATCH for Cedar authorization only; self-service updates are forbidden (403). Not persisted on PATCH."
          },
          "display_code": { "type": "string", "nullable": true },
          "first_name": { "type": "string", "nullable": true },
          "last_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "roles": { "type": "array", "items": { "type": "string" } },
          "tos_accepted": {
            "type": "boolean",
            "enum": [true],
            "description": "Self-service only; may only be set to true to record TOS acceptance"
          }
        },
        "additionalProperties": false
      },
      "UserCreateRequest": {
        "type": "object",
        "properties": {
          "tenant_uuid": {
            "type": "string",
            "format": "uuid",
            "description": "Tenant that owns the new user row"
          },
          "display_code": { "type": "string", "nullable": true },
          "uuid": {
            "type": "string",
            "format": "uuid",
            "description": "Optional stable user id; upserts when the row already exists (operator seeding only)"
          },
          "idp_id": {
            "type": "string",
            "description": "Optional identity provider subject; defaults to enrolled_{uuid} when omitted"
          },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "roles": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1,
            "description": "Role slugs for the new user (e.g. patient.self for enrollment)"
          }
        },
        "required": ["tenant_uuid", "first_name", "last_name", "email", "roles"],
        "additionalProperties": false
      },
      "UserProvisionRequest": {
        "type": "object",
        "properties": {
          "tenant_uuid": { "type": "string", "format": "uuid" },
          "idp_id": { "type": "string", "minLength": 1 },
          "display_code": { "type": "string", "nullable": true },
          "first_name": { "type": "string", "nullable": true },
          "last_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true }
        },
        "required": ["tenant_uuid", "idp_id", "first_name", "last_name", "email"],
        "additionalProperties": false
      },
      "UserAuditItem": {
        "type": "object",
        "properties": {
          "history_uuid": { "type": "string", "format": "uuid", "nullable": true },
          "uuid": { "type": "string", "format": "uuid" },
          "tenant_uuid": { "type": "string", "format": "uuid" },
          "idp_id": { "type": "string", "minLength": 1 },
          "display_code": { "type": "string", "nullable": true },
          "first_name": { "type": "string", "nullable": true },
          "last_name": { "type": "string", "nullable": true },
          "email": { "type": "string", "nullable": true },
          "roles": { "type": "array", "items": { "type": "string" } },
          "changed_at": { "type": "string", "format": "date-time" },
          "changed_by_uuid": { "type": "string", "format": "uuid" },
          "changed_by_type": { "type": "integer" },
          "change_type": { "type": "integer" }
        },
        "required": ["uuid", "tenant_uuid", "idp_id", "roles", "changed_at", "changed_by_uuid", "changed_by_type", "change_type"]
      },
      "UserAuditListResponse": {
        "type": "object",
        "properties": {
          "user_uuid": { "type": "string", "format": "uuid" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/UserAuditItem" } },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" }
        },
        "required": ["user_uuid", "items", "total", "page", "page_size"]
      },
      "RoleCatalog": {
        "type": "object",
        "properties": {
          "actor_classes": {
            "type": "array",
            "items": { "type": "string" },
            "description": "Platform tier-1 actors (operator, study, clinician, patient); vocabulary, not caller JWT"
          },
          "roles": { "type": "array", "items": { "type": "string" } },
          "role_definitions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string" },
                "label": { "type": "string" }
              },
              "required": ["id", "label"]
            }
          },
          "tenant_types": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": { "type": "string" }
            }
          }
        },
        "required": ["actor_classes", "roles", "role_definitions", "tenant_types"]
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "invalid_request",
              "conflict",
              "not_found",
              "forbidden",
              "unauthenticated",
              "unauthorized",
              "authorization_unavailable",
              "database error",
              "internal_server_error",
              "http_error",
              "method_not_allowed",
              "payload_too_large"
            ]
          },
          "message": { "type": "string" }
        }
      }
    }
  }
}